+++ spec/app/create_broker_spec.rb
@@ -18,7 +18,7 @@ describe Razor::Command::CreateBroker do
       header 'content-type', 'application/json'
 
       Razor.config['broker_path'] =
-        (Pathname(__FILE__).dirname.parent + 'fixtures' + 'brokers').realpath.to_s
+          (Pathname(__FILE__).dirname.parent + 'fixtures' + 'brokers').realpath.to_s
     end
 
     let :command_hash do
@@ -30,7 +30,13 @@ describe Razor::Command::CreateBroker do
     it_behaves_like "a command"
 
     def create_broker(params)
-      command 'create-broker', params
+      # These configuration parameters are added to the return response from
+      # the server. Our validation needs a way to anticipate these in the
+      # response, so they're passed to the `command` helper.
+      expect = if params['broker_type'] == 'with_configuration'
+                 {'configuration' => {'key-with-default' => 1, 'optional-key-with-default' => 1}}
+               end || {}
+      command 'create-broker', params, expect: expect
       params
     end
 
@@ -52,7 +58,7 @@ describe Razor::Command::CreateBroker do
 
       last_response.status.should == 404
       last_response.json['error'].should ==
-        "broker_type must be the name of an existing broker type, but is 'no-such-broker-for-me'"
+          "broker_type must be the name of an existing broker type, but is 'no-such-broker-for-me'"
     end
 
     it "should fail cleanly if 'configuration' is a string" do
@@ -79,8 +85,15 @@ describe Razor::Command::CreateBroker do
 
       Razor::Data::Broker[:name => command['name']].should be_an_instance_of Razor::Data::Broker
     end
-    # Successful creation
+
     it "should return 202 when repeated with the same parameters" do
+      create_broker command_hash
+      create_broker command_hash
+
+      last_response.status.should == 202
+    end
+
+    it "should return 409 when repeated with slightly different parameters" do
       command = create_broker command_hash
       command_hash['name'] = command_hash['name'].upcase
       command = create_broker command_hash
@@ -91,16 +104,16 @@ describe Razor::Command::CreateBroker do
 
     it "should validate valid configuration" do
       command_hash['broker_type'] = 'with_configuration'
-      command_hash['configuration'] = {'some-key' => 'valid-value'}
+      command_hash['configuration'] = {'required-key' => 'valid-value'}
       command = create_broker command_hash
 
       last_response.status.should == 202
-      Razor::Data::Broker[:name => command['name']].configuration['some-key'].should == 'valid-value'
+      Razor::Data::Broker[:name => command['name']].configuration['required-key'].should == 'valid-value'
     end
 
     it "should validate invalid configuration" do
       command_hash['broker_type'] = 'with_configuration'
-      command_hash['configuration'] = {'not-valid-key' => 'not-valid-value'}
+      command_hash['configuration'] = {'not-valid-key' => 'not-valid-value', 'required-key' => 'value'}
       command = create_broker command_hash
 
       last_response.json['error'].should == "configuration key 'not-valid-key' is not defined for this broker type"
@@ -109,16 +122,16 @@ describe Razor::Command::CreateBroker do
 
     it "should validate valid configuration abbreviation" do
       command_hash['broker_type'] = 'with_configuration'
-      command_hash['c'] = {'some-key' => 'valid-value'}
+      command_hash['c'] = {'required-key' => 'valid-value'}
       command = create_broker command_hash
 
       last_response.status.should == 202
-      Razor::Data::Broker[:name => command['name']].configuration['some-key'].should == 'valid-value'
+      Razor::Data::Broker[:name => command['name']].configuration['required-key'].should == 'valid-value'
     end
 
     it "should validate invalid configuration abbreviation" do
       command_hash['broker_type'] = 'with_configuration'
-      command_hash['c'] = {'not-valid-key' => 'not-valid-value'}
+      command_hash['c'] = {'not-valid-key' => 'not-valid-value', 'required-key' => 'value'}
       command = create_broker command_hash
 
       last_response.json['error'].should == "configuration key 'not-valid-key' is not defined for this broker type"
@@ -127,14 +140,14 @@ describe Razor::Command::CreateBroker do
 
     it "should validate mixed shorthand and longhand configuration" do
       command_hash['broker_type'] = 'with_configuration'
-      command_hash['c'] = {'some-key' => 'valid-value'}
-      command_hash['configuration'] = {'some-other-key' => 'other-valid-value'}
+      command_hash['c'] = {'required-key' => 'valid-value'}
+      command_hash['configuration'] = {'key-with-default' => 'other-valid-value'}
       command = create_broker command_hash
 
       last_response.status.should == 202
       configuration = Razor::Data::Broker[:name => command['name']].configuration
-      configuration['some-key'].should == 'valid-value'
-      configuration['some-other-key'].should == 'other-valid-value'
+      configuration['required-key'].should == 'valid-value'
+      configuration['key-with-default'].should == 'other-valid-value'
     end
   end
 end